requirements.txt
requests
colorama
discord.py-self
gtts
qrcode
pyfiglet
aiohttp
Pillow
